Adding webIDL sources to preferences
In this branch properties of WebIDL files can be marked as taint
sources with an attribute. This is super nice, but if one wants to
disable one of them one has to note down their names during a build.
This is pretty cumbersome. This commit tries to resolve this issue,
as it simply adds every IDL based source to the properties
defaults.